+++ structurizr-dsl/src/main/java/com/structurizr/dsl/CustomElementParser.java
@@ -13,7 +13,7 @@ final class CustomElementParser extends AbstractParser {
     private final static int DESCRIPTION_INDEX = 3;
     private final static int TAGS_INDEX = 4;
 
-    CustomElement parse(ModelDslContext context, Tokens tokens) {
+    CustomElement parse(ModelDslContext context, Tokens tokens, Archetype archetype) {
         // element <name> [metadata] [description] [tags]
 
         if (tokens.hasMoreThan(TAGS_INDEX)) {
@@ -26,22 +26,23 @@ final class CustomElementParser extends AbstractParser {
 
         String name = tokens.get(NAME_INDEX);
 
-        String metadata = "";
+        String metadata = archetype.getMetadata();
         if (tokens.includes(METADATA_INDEX)) {
             metadata = tokens.get(METADATA_INDEX);
         }
 
-        String description = "";
+        String description = archetype.getDescription();
         if (tokens.includes(DESCRIPTION_INDEX)) {
             description = tokens.get(DESCRIPTION_INDEX);
         }
 
         CustomElement customElement = context.getWorkspace().getModel().addCustomElement(name, metadata, description);
 
+        String[] tags = archetype.getTags().toArray(new String[0]);
         if (tokens.includes(TAGS_INDEX)) {
-            String tags = tokens.get(TAGS_INDEX);
-            customElement.addTags(tags.split(","));
+            tags = tokens.get(TAGS_INDEX).split(",");
         }
+        customElement.addTags(tags);
 
         if (context.hasGroup()) {
             customElement.setGroup(context.getGroup().getName());